+++ tests/Method/SendDiceTest.php
@@ -41,6 +41,7 @@ final class SendDiceTest extends TestCase
             'meid1',
             $replyParameters,
             $replyMarkup,
+            true,
         );
 
         $this->assertSame(
@@ -51,6 +52,7 @@ final class SendDiceTest extends TestCase
                 'emoji' => 'emoji1',
                 'disable_notification' => true,
                 'protect_content' => false,
+                'allow_paid_broadcast' => true,
                 'message_effect_id' => 'meid1',
                 'reply_parameters' => $replyParameters->toRequestArray(),
                 'reply_markup' => $replyMarkup->toRequestArray(),