+++ tests/Method/SendMessageTest.php
@@ -49,6 +49,7 @@ final class SendMessageTest extends TestCase
             'meid1',
             $replyParameters,
             $replyMarkup,
+            true,
         );
 
         $this->assertSame(
@@ -62,6 +63,7 @@ final class SendMessageTest extends TestCase
                 'link_preview_options' => $linkPreviewOptions->toRequestArray(),
                 'disable_notification' => true,
                 'protect_content' => false,
+                'allow_paid_broadcast' => true,
                 'message_effect_id' => 'meid1',
                 'reply_parameters' => $replyParameters->toRequestArray(),
                 'reply_markup' => $replyMarkup->toRequestArray(),