+++ tests/Method/SendVoiceTest.php
@@ -51,6 +51,7 @@ final class SendVoiceTest extends TestCase
             'meID',
             $replyParameters,
             $replyMarkup,
+            true,
         );
 
         $this->assertSame(
@@ -65,6 +66,7 @@ final class SendVoiceTest extends TestCase
                 'duration' => 56,
                 'disable_notification' => true,
                 'protect_content' => false,
+                'allow_paid_broadcast' => true,
                 'message_effect_id' => 'meID',
                 'reply_parameters' => $replyParameters->toRequestArray(),
                 'reply_markup' => $replyMarkup->toRequestArray(),