+++ tests/Method/Sticker/SendStickerTest.php
@@ -46,6 +46,7 @@ final class SendStickerTest extends TestCase
             'meid1',
             $replyParameters,
             $replyMarkup,
+            true,
         );
 
         $this->assertSame(
@@ -57,6 +58,7 @@ final class SendStickerTest extends TestCase
                 'emoji' => '👍',
                 'disable_notification' => true,
                 'protect_content' => false,
+                'allow_paid_broadcast' => true,
                 'message_effect_id' => 'meid1',
                 'reply_parameters' => $replyParameters->toRequestArray(),
                 'reply_markup' => $replyMarkup->toRequestArray(),